+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Statue.java
@@ -151,8 +151,10 @@ public class Statue extends Mob {
 	
 	@Override
 	public void die( Object cause ) {
-		weapon.identify(false);
-		Dungeon.level.drop( weapon, pos ).sprite.drop();
+		if (!Dungeon.alchemist) {
+			weapon.identify(false);
+			Dungeon.level.drop(weapon, pos).sprite.drop();
+		}
 		super.die( cause );
 	}
